Reorganized algorithm modules
